scripts/storm-ads/requirements.txt
google-ads>=25.0.0
google-auth>=2.0.0
requests>=2.28.0
